requirements.txt
flask==3.1.0
opencv-python==4.11.0.86
numpy==1.26.4
mediapipe==0.10.21
joblib==1.4.2
stanza==1.11.0
flask-mail==0.10.0
pymongo==4.11.1
werkzeug==3.1.3
python-dotenv==1.0.1
flask-cors==5.0.1
NumPy==1.26.4
TensorFlow==2.18.0
MediaPipe==0.10.21